# wismap/openapi.yaml
openapi: 3.1.0
info:
  title: WisMAP API
  version: "0.5.1"      # MUST match wismap.__version__ — checked by tests/check_openapi_coverage.py
  description: |
    Versioned JSON API for the RAKwireless WisBlock pin-mapping tool.

    This document is the **canonical machine-readable contract**. The
    rationale and prose narrative live in the spec workflow under
    `.sdd/specs/005-wismap-api-v1/` (consumer team's draft + our accepted
    deviations). Read those before negotiating any breaking change.

    ## Authentication

    `POST /validate` and `POST /solve` are gated (spec 009-api-key-auth): send
    `Authorization: Bearer <key>`. All discovery/read endpoints stay public. The
    browser SPA uses a session+CSRF cookie pair instead of a key.

    ## Conventions

    Deviations from the consumer team's Appendix A sketch are noted with
    references into §`N` of `.sdd/specs/005-wismap-api-v1/review.md`.

    - **Slot names** are descriptive (`SENSOR_A`, `IO_A`, `CORE`, `POWER`),
      not single-letter. (§3.1)
    - **Pin role names** are unprefixed (`I2C1_SDA`). The Arduino-style
      `WB_*` alias is exposed alongside in
      `resolved.slots[*].pins[*].wisblock_pin`. (§3.2)
    - **Module taxonomy** exposes both `type` (WisMAP physical class:
      `WisCore | WisIO | WisSensor | WisPower | WisModule | Accessories`)
      and `category` (functional:
      `sensor | io | display | communication | storage | power | core | base`).
      (§3.3)
...

    ## Extension fields

...

...

...
servers:
...
paths:
...
  /api/v1/cores/{id}:
    get:
      operationId: getCore
      summary: Get one Core
      tags: [Cores]
      parameters:
        - $ref: "#/components/parameters/CoreIdPath"
        - $ref: "#/components/parameters/ShowNc"
      responses:
        "200":
          description: Core detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Core" }
              example:
                id: RAK4631
                name: RAK4631 nRF52840 LoRaWAN
                mcu: Nordic nRF52840
                lora_chip: Semtech SX1262
                peripherals: [GPIO, I2C1, I2C2, SPI1, UART0, UART1, analog]
                lora_pins: null
                power_pins:
                  3V3_S_control: P1.02
                compatible_bases: [RAK19001, RAK19003, RAK19007, RAK19009, RAK19010, RAK19011, RAK5005-O]
                notes: []
                datasheet_url: https://docs.rakwireless.com/product-categories/wisblock/rak4631/overview/
                pin_mapping:
                  - { pin: "10", function: "P1.01" }
                  - { pin: "11", function: "P0.20" }
                images: []
                schematics: []
                tags: [lorawan, lora]
        "404":
          $ref: "#/components/responses/CoreNotFound"

  # ─── Bases ──────────────────────────────────────────────────────────────
...
  /api/v1/bases/{id}:
    get:
      operationId: getBase
      summary: Get one Base
      tags: [Bases]
      parameters:
        - $ref: "#/components/parameters/BaseIdPath"
        - $ref: "#/components/parameters/ShowNc"
      responses:
        "200":
          description: Base detail with per-slot pin geometry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Base" }
              example:
                id: RAK19007
                name: RAK19007 Base Board 2nd Gen
                form_factor: normal
                slots: [CORE, IO_A, SENSOR_A, SENSOR_B, SENSOR_C, SENSOR_D]
                core_socket: WB_CORE_S1
                slot_pin_map:
                  SENSOR_A:
                    I2C1_SCL: I2C1_SCL
                    I2C1_SDA: I2C1_SDA
                    SPI_CS: SPI_CS
                shared_buses:
                  - { bus: I2C1, slots: [CORE, IO_A, SENSOR_A, SENSOR_B, SENSOR_C, SENSOR_D] }
                datasheet_url: https://docs.rakwireless.com/product-categories/wisblock/rak19007/overview/
                slot_info:
                  CORE: { double: false, double_blocks: null, accepts_type: WisCore, layer: top }
                  IO_A: { double: false, double_blocks: null, accepts_type: WisIO, layer: top }
                  SENSOR_A: { double: true, double_blocks: SENSOR_B, accepts_type: WisSensor, layer: top }
                  SENSOR_C: { double: false, double_blocks: null, accepts_type: WisSensor, layer: bottom }
                notes:
                  - "Accepts long sensor on SENSOR_A slot but blocking SENSOR_B slot"
        "404":
          $ref: "#/components/responses/BaseNotFound"

  # ─── Modules ────────────────────────────────────────────────────────────
  /api/v1/modules:
    get:
      operationId: listModules
      summary: List modules
      tags: [Modules]
      description: |
        WisIO, WisSensor, WisModule, WisPower, and Accessories. Cores and
        Bases have their own endpoints.
      parameters:
        - $ref: "#/components/parameters/Type"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/Interface"
        - $ref: "#/components/parameters/CompatibleWithCore"
      responses:
        "200":
          description: Modules in summary form, filtered by the query parameters.
          content:
            application/json:
              schema:
                type: object
                required: [modules]
                properties:
                  modules:
                    type: array
                    items: { $ref: "#/components/schemas/ModuleSummary" }
              example:
                modules:
                  - id: RAK1901
                    name: RAK1901 Temperature and Humidity Sensor
                    chip: Sensirion SHTC3
                    type: WisSensor
                    category: sensor
                    interfaces: [I2C]
                    i2c_addresses: ["0x70"]
                    compatible_slots:
                      RAK19007: [SENSOR_A, SENSOR_B, SENSOR_C, SENSOR_D]
                    datasheet_url: https://docs.rakwireless.com/product-categories/wisblock/rak1901/overview/
                    tags: [temperature, humidity, environmental, i2c]
                    double: false
  /api/v1/modules/{id}:
    get:
      operationId: getModule
      summary: Get one module
      tags: [Modules]
      parameters:
        - $ref: "#/components/parameters/ModuleIdPath"
        - $ref: "#/components/parameters/ShowNc"
      responses:
        "200":
          description: Module detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Module" }
              example:
                id: RAK1901
                name: RAK1901 Temperature and Humidity Sensor
                chip: Sensirion SHTC3
                type: WisSensor
                category: sensor
                interfaces: [I2C]
                i2c_addresses: ["0x70"]
                i2c_alternate_addresses: []
                signal_map:
                  SDA: { role: I2C_SDA }
                  SCL: { role: I2C_SCL }
                compatible_slots:
                  RAK19007: [SENSOR_A, SENSOR_B, SENSOR_C, SENSOR_D]
                core_requirements:
                  peripherals: [I2C1]
                power:
                  rail: 3V3
                  control_pin: null
                  warmup_ms: null
                  current_typical_uA: null
                  current_peak_mA: null
                electrical_notes: null
                datasheet_url: https://docs.rakwireless.com/product-categories/wisblock/rak1901/overview/
                example_urls: []
                tags: [temperature, humidity, environmental, i2c]
                images: []
                schematics: []
                pin_mapping:
                  - { pin: "2", function: GND }
                  - { pin: "7", function: I2C_SCL }
                  - { pin: "8", function: I2C_SDA }
                double: false
                notes: []
        "404":
          $ref: "#/components/responses/ModuleNotFound"

  # ─── Validate ───────────────────────────────────────────────────────────
  /api/v1/validate:
    post:
      operationId: validate
      summary: Validate a configuration
      tags: [Validation]
      security:
        - bearerAuth: []
      description: |
        Returns 200 with `valid: true|false` for every well-formed request.
        Conflicts are **not errors** — they appear in the `conflicts[]`
        array. The 4xx responses are reserved for problems with the request
        itself (malformed JSON, duplicate slot, unknown core/base, etc.).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ValidateRequest" }
            examples:
              valid-clean:
                summary: Clean valid config (fixture 01)
                value:
                  core: RAK4631
                  base: RAK19007
                  slots:
                    - { slot: SENSOR_A, module: RAK1901 }
                    - { slot: SENSOR_B, module: RAK1902 }
              i2c-collision:
                summary: Same module twice — I2C address collision (fixture 02)
                value:
                  core: RAK4631
                  base: RAK19007
                  slots:
                    - { slot: SENSOR_A, module: RAK1901 }
                    - { slot: SENSOR_C, module: RAK1901 }
              i2c-override:
                summary: Collision resolved by jumper override (fixture 03)
                value:
                  core: RAK4631
                  base: RAK19007
                  slots:
                    - { slot: SENSOR_A, module: RAK1901 }
                    - { slot: SENSOR_C, module: RAK1901 }
                  options:
                    i2c_address_overrides:
                      RAK1901@SENSOR_C: "0x71"
              coreless-base:
                summary: Coreless base (RAK6421 Pi Hat), core omitted (fixture 14)
                value:
                  base: RAK6421
                  slots:
                    - { slot: SENSOR_A, module: RAK1901 }
      responses:
        "200":
          description: |
            Validation completed. Check `valid` and the `conflicts[]` /
            `warnings[]` arrays. `resolved` is null when `valid` is false.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValidateResponse" }
              examples:
                valid:
                  summary: "Clean config → valid:true, full resolved block"
                  value:
                    valid: true
                    conflicts: []
                    warnings: []
                    resolved:
                      core: RAK4631
                      base: RAK19007
                      slots:
                        SENSOR_A:
                          module: RAK1901
                          i2c_address: "0x70"
                          pins:
                            SDA: { role: I2C1_SDA, wisblock_pin: WB_I2C1_SDA, mcu_pin: P0.13 }
                            SCL: { role: I2C1_SCL, wisblock_pin: WB_I2C1_SCL, mcu_pin: P0.14 }
                          power: { rail: 3V3, control_pin: null, warmup_ms: null, current_typical_uA: null, current_peak_mA: null }
                      buses:
                        I2C1:
                          slots: [CORE, SENSOR_A, SENSOR_B]
                          devices:
                            - { module: RAK1901, slot: SENSOR_A, address: "0x70" }
                      lorawan:
                        chip: Semtech SX1262
                        spi: SPI1
                        pins: null
                conflict:
                  summary: "I2C address collision → valid:false, structured conflict"
                  value:
                    valid: false
                    conflicts:
                      - code: i2c_address_collision
                        message: "Possible conflict with I2C addresses"
                        severity: error
                        involves:
                          - { module: RAK1901, slot: SENSOR_A }
                          - { module: RAK1901, slot: SENSOR_C }
                        context:
                          function: I2C_ADDR
                          addresses: ["0x70"]
                          bus: I2C1
                    warnings: []
                    resolved: null
        "400":
          description: Malformed request body (bad JSON, missing required field, wrong type).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: invalid_request
                  message: "Field `core` is required for this base."
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Unknown core or base id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: core_not_found
                  message: "Core 'RAK9999' is not known to WisMAP."
        "422":
          description: |
            Request shape valid but semantically rejectable: duplicate slot
            in `slots[]`, or `CORE` entry in `slots[]` that disagrees with
            the top-level `core`. See §3.5 of the consumer-contract review.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: duplicate_slot
                  message: "Slot 'SENSOR_A' appears more than once."
  /api/v1/solve:
    post:
      operationId: solve
      summary: Solve module placement (which slot each module goes in)
      tags: [Validation]
      security:
        - bearerAuth: []
      description: |
        Given a `core` + `base` + a flat list of desired `modules`, return up to
        `max_solutions` (default 3, max 5) ranked slot assignments. Answers
        *where modules go* — placements + scores only, **no `resolved` pin map**.
        The client calls `POST /api/v1/validate` on the layout it picks to get
        pins / buses / lorawan. Solutions rank by most-placed → fewest errors →
        fewest warnings → most sensors on the top layer → deterministic slot
        order. Unknown / incompatible / over-capacity modules appear in each
        solution's `unplaced[]` (they do not 404 the request).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SolveRequest" }
            examples:
              two-sensors:
                summary: Two sensors on a standard base
                value:
                  core: RAK4631
                  base: RAK19007
                  modules: [RAK1901, RAK1902]
              over-capacity:
                summary: More sensors than slots — best partial, top 5
                value:
                  core: RAK4631
                  base: RAK19007
                  modules: [RAK1901, RAK1902, RAK1903, RAK1904, RAK12500]
                  max_solutions: 5
      responses:
        "200":
          description: |
            Ranked placements. `solution_count` ≤ `max_solutions`; `truncated`
            is true only when the search cap tripped (results then best-effort).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SolveResponse" }
              example:
                core: RAK4631
                base: RAK19007
                requested_modules: [RAK1901, RAK1902]
                solution_count: 2
                truncated: false
                solutions:
                  - rank: 1
                    valid: true
                    error_count: 0
                    warning_count: 0
                    conflicts: []
                    warnings: []
                    sensors_on_top: 2
                    slots:
                      - { slot: SENSOR_A, module: RAK1901 }
                      - { slot: SENSOR_B, module: RAK1902 }
                    unplaced: []
                warnings: []
        "400":
          description: Malformed request body (bad JSON, missing/mistyped field).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: invalid_request
                  message: "Field `modules` must be an array."
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Unknown core or base id.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error:
                  code: base_not_found
                  message: "Base 'RAK9999' is not known to WisMAP."

# ============================================================================
# Components
# ============================================================================
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Opaque API key sent as `Authorization: Bearer <key>`. Required on
        `POST /validate` and `POST /solve`. Keys are issued and rotated
        server-side (see spec 009-api-key-auth). The browser SPA authenticates
        with a session+CSRF cookie pair instead and does not send this header.

  # ─── Parameters ─────────────────────────────────────────────────────────
  parameters:
    CoreIdPath:
      name: id
      in: path
      required: true
      description: Core id (case-insensitive on input; uppercase on output).
      schema: { type: string, examples: [RAK4631] }
    BaseIdPath:
      name: id
      in: path
      required: true
      description: Base id (case-insensitive on input; uppercase on output).
      schema: { type: string, examples: [RAK19007] }
    ModuleIdPath:
      name: id
      in: path
      required: true
      description: Module id (case-insensitive on input; uppercase on output).
      schema: { type: string, examples: [RAK1901] }
    ShowNc:
      name: show_nc
      in: query
      required: false
      description: |
        When `true`, `pin_mapping` and `pin_mapping_table` include NC
        (not-connected) pins. Default `false` for compact responses.
      schema: { type: boolean, default: false }
    Type:
      name: type
      in: query
      required: false
      description: Filter by WisMAP physical class.
      schema:
        type: string
        enum: [WisIO, WisSensor, WisModule, WisPower, Accessories]
    Category:
      name: category
      in: query
      required: false
      description: Filter by functional category.
      schema:
        type: string
        enum: [sensor, io, display, communication, storage, power, core, base]
    Interface:
      name: interface
      in: query
      required: false
      description: Filter by bus family the module uses.
      schema:
        type: string
        enum: [I2C, SPI, UART, GPIO, analog, I2S, PDM]
    CompatibleWithCore:
      name: compatible_with_core
      in: query
      required: false
      description: |
        Filter to modules compatible with the named Core. Accepted in any case.
      schema: { type: string, examples: [RAK4631] }

  # ─── Responses ──────────────────────────────────────────────────────────
  responses:
    CoreNotFound:
      description: Core id not in WisMAP catalog.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: core_not_found
              message: "Core 'RAK9999' is not known to WisMAP."
              details: { core: "RAK9999" }
    BaseNotFound:
      description: Base id not in WisMAP catalog.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: base_not_found
              message: "Base 'RAK9999' is not known to WisMAP."
              details: { base: "RAK9999" }
    ModuleNotFound:
      description: Module id not in WisMAP catalog (and not a Core or Base).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: module_not_found
              message: "Module 'RAK9999' is not known to WisMAP."
              details: { module: "RAK9999" }
    Forbidden:
      description: |
        Authentication required/failed. `POST /validate` and `POST /solve` need
        either a valid `Authorization: Bearer` API key or the browser SPA's
        session+CSRF pair. See spec 009-api-key-auth.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: forbidden
              message: "Valid API key required"

  # ─── Schemas ────────────────────────────────────────────────────────────
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                One of: `invalid_request`, `unauthorized`, `forbidden`,
                `core_not_found`, `base_not_found`, `module_not_found`,
                `duplicate_slot`, `unprocessable`, `rate_limited`,
                `internal_error`, `service_unavailable`.
            message: { type: string }
            details:
              type: object
              additionalProperties: true

    # — Health —
...
    Core:
      allOf:
        - $ref: "#/components/schemas/CoreSummary"
        - type: object
          properties:
            lora_pins:
              type: [object, "null"]
              description: |
                MCU-pin map for the on-board LoRa SPI. `null` in v1 until
                Phase 6 of the data-backfill spec populates them.
              additionalProperties: { type: string }
            power_pins:
              type: [object, "null"]
              description: |
                MCU pins for power-rail control. Currently only
                `3V3_S_control` is populated.
              additionalProperties: { type: string }
            compatible_bases:
              type: array
              items: { type: string }
              description: Bases that expose a CORE slot.
            notes:
              type: array
              items: { type: string }
            pin_mapping:
              type: array
              items: { $ref: "#/components/schemas/PinMappingEntry" }
              description: |
                **Convenience field; consumers may ignore.** Per-pin function
                map of the Core (pin number → MCU pin or signal name) used by
                the WisMAP frontend's ModuleDetail view.
            images:
              type: array
              items: { type: string, format: uri }
            schematics:
              type: array
              items: { type: string, format: uri }

    # — Bases —
...
    Base:
      allOf:
        - $ref: "#/components/schemas/BaseSummary"
        - type: object
          properties:
            slot_pin_map:
              type: object
              description: |
                Per-slot pin-role map. Key is the slot name, value is a dict
                where both keys and values are role names. Example:
                `{ "SENSOR_A": { "I2C1_SDA": "I2C1_SDA", "SPI_CS": "SPI_CS" } }`.
              additionalProperties:
                type: object
                additionalProperties: { type: string }
            shared_buses:
              type: array
              description: Which slots share each bus instance.
              items:
                type: object
                properties:
                  bus:   { type: string }
                  slots:
                    type: array
                    items: { type: string }
            slot_info:
              type: object
              description: |
                **Convenience field; consumers may ignore.** Per-slot
                double/blocking/accepts-type metadata used by the frontend's
                combine-tool dropdowns.
              additionalProperties: { $ref: "#/components/schemas/SlotInfo" }
            pin_mapping_table:
              type: [object, "null"]
              description: |
                **Convenience field; consumers may ignore.** Wide-format
                pin-by-slot table used by the frontend's ModuleDetail view.
              properties:
                columns:
                  type: array
                  items: { type: string }
                rows:
                  type: array
                  items:
                    type: object
                    additionalProperties: { type: string }
            notes:
              type: array
              items: { type: string }
            images:
              type: array
              items: { type: string, format: uri }
            schematics:
              type: array
              items: { type: string, format: uri }
            i2c_address:
              type: [string, "null"]
              description: |
                Some bases (e.g. RAK6421 Pi Hat) embed their own I2C device.
                When present, this is its address.
            naming:
              type: object
              description: Per-base role-name overrides (legacy field).
              additionalProperties: { type: string }
...
    ModuleSummary:
      type: object
      required: [id, name, type]
      properties:
        id:   { type: string, examples: [RAK1901] }
        name: { type: string }
        chip:
          type: [string, "null"]
          description: |
            Manufacturer-prefixed primary chip name where known
            (e.g. `Sensirion SHTC3`). Many modules are not yet populated;
            see `.sdd/specs/007-power-data-backfill/`.
        type:
          type: string
          enum: [WisIO, WisSensor, WisModule, WisPower, Accessories]
        category:
          type: [string, "null"]
          enum: [sensor, io, display, communication, storage, power, null]
        interfaces:
          type: array
          description: Bus families used by this module.
          items:
            type: string
            enum: [I2C, SPI, UART, GPIO, analog, I2S, PDM]
        i2c_addresses:
          type: array
          items: { type: string, examples: ["0x70"] }
        compatible_slots:
          type: object
          description: |
            For each base id, the list of slot ids on that base where this
            module fits. Slot-type compatibility and double-sensor blocking
            are pre-resolved here.
          additionalProperties:
            type: array
            items: { type: string }
        datasheet_url:
          type: [string, "null"]
          format: uri
        tags:
          type: array
          items: { type: string }
        double:
          type: boolean
          description: |
            True for double-size sensors (occupy two adjacent sensor slots).
            Always false for non-WisSensor types.
...